requirements.txt
langgraph
groq
python-dotenv
requests
biopython
